feat: 구입 금액에 해당 하는 만큼 로또 발행 구현

+++ src/main/java/lotto/view/InputView.java
@@ -4,8 +4,8 @@ import camp.nextstep.edu.missionutils.Console;
 
 public class InputView {
 
-    public Long inputPurchaseAmount() {
+    public int inputPurchaseAmount() {
         String userInput = Console.readLine();
-        return Long.parseLong(userInput);
+        return Integer.parseInt(userInput);
     }
 }
